webapp/streamlit/requirements.txt
deta
python-dotenv
streamlit
absl-py==1.4.0
alabaster==0.7.13
altair==4.2.2
ansicon==1.89.0
anyio==3.6.2
asgiref==3.6.0
astroid==2.15.4
asttokens==2.0.8
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.1
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
black==22.10.0
bleach==5.0.1
blessed==1.20.0
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
deta==1.1.0
dill==0.3.6
distlib==0.3.6
Django==4.1.7
django-crispy-forms==2.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
extra-streamlit-components==0.1.56
filelock==3.9.0
firebase-admin==6.1.0
Flask==2.2.3
fonttools==4.38.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.51.1
grpcio-status==1.48.2
h11==0.14.0
h2o==3.40.0.4
h2o-wave==0.25.2
h2o-wave-university==0.25.2
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
idna==3.3
imagesize==1.4.1
importlib-metadata==5.2.0
inquirer==3.1.3
ipykernel==6.16.0
ipython==8.5.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.1
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
jsonschema==4.17.3
jupyter-core==4.11.1
jupyter_client==7.3.5
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.0.0
MouseInfo==0.1.3
mpmath==1.2.1
msgpack==1.0.5
music==0.7b3
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathspec==0.10.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.13
pkginfo==1.9.2
platformdirs==2.5.2
plotly==5.14.1
plotly-express==0.4.1
pprintpp==0.4.0
prompt-toolkit==3.0.31
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
pyflowchart==0.2.3
PyGetWindow==0.0.9
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.17.3
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.9
PyPDF2==2.11.1
PyPDF4==1.27.0
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-etcd==0.4.5
PythonTurtle==0.3.2
pytweening==1.0.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
readchar==4.0.5
readme-renderer==37.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.0.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.0
seaborn==0.12.2
ShowCallStack==0.3.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
SpeechRecognition==3.8.1
sqlparse==0.4.3
stack-data==0.5.1
starlette==0.26.1
statsmodels==0.13.5
streamlit==1.22.0
streamlit-authenticator==0.2.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
twine==4.0.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.22.0
validators==0.20.0
virtualenv==20.19.0
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.15.0
yarg==0.1.9
zipp==3.11.0